+++ projects/u7chat/src/app/page.tsx
@@ -1,4 +1,4 @@
-import { Chat } from '@/features/chat/chat'
+import { ChatMessage } from '@/features/chat-message/chat-message'
 import { fetchModels } from '@/features/llm-models/llm-model-actions'
 import { ModelSelect } from '@/features/llm-models/model-select'
 
@@ -10,7 +10,7 @@ export default async function Page(props: PageProps<'/'>) {
     <div className='flex min-h-screen flex-col items-center justify-center p-4'>
       <div className='w-full max-w-lg rounded border border-gray-200 p-4'>
         <ModelSelect models={models} defaultModel={selectModel} />
-        {selectModel && <Chat model={selectModel} />}
+        {selectModel && <ChatMessage model={selectModel} />}
       </div>
     </div>
   )
